Add more unit tests
