+++ dynamic-simulation/dynamic-simulation-api/src/test/resources/network.xiidm
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<iidm:network xmlns:iidm="http://www.itesla_project.eu/schema/iidm/1_0" id="network" caseDate="2019-11-07T10:30:00.000+01:00" forecastDistance="0" sourceFormat="test">
+</iidm:network>
